+++ domain/src/main/java/ru/dpopkov/knownics/domain/question/Category.java
@@ -7,6 +7,14 @@ public class Category extends ModifiableEntity {
     private String name;
     private String description;
 
+    public Category() {
+    }
+
+    public Category(String name, String description) {
+        this.name = name;
+        this.description = description;
+    }
+
     public String getName() {
         return name;
     }