+++ services/src/main/java/ru/dpopkov/knownics/services/CategoryService.java
@@ -1,0 +1,6 @@
+package ru.dpopkov.knownics.services;
+
+import ru.dpopkov.knownics.domain.question.Category;
+
+public interface CategoryService extends BaseService<Category> {
+}